lab/pi/ssd1306_eyes/requirements.txt
# Install on the Raspberry Pi (venv or --user). Dev-Host pytest uses PIL only for dirty_stats tests.
luma.oled>=3.13,<4
pillow>=10,<12
